refactor(participant-portal): extract ScoreEventsTable from ScoreEvents.tsx
ScoreEvents.tsx rendered both the cumulative-score chart and the score-history
table in one page, pulling 17 imports (over the high-coupling threshold of 16).
The history table owned the only Table/Badge uses, the lib/format time helpers,
and the SOURCE_KEY/SOURCE_COLOR maps. Extracting it to ScoreEventsTable.tsx
drops ScoreEvents.tsx to 15 imports (off the high-coupling list) and gives the
table its own single-responsibility module.

Relates #1418.

## Regression analysis
- Pure structural move: the table JSX + its SOURCE maps are byte-identical, only
  relocated and rendered via <ScoreEventsTable entries={data.entries} />.
  typecheck passes; all 274 participant-portal tests pass unchanged.

## Physical impact
- NO-OP on CFn / deployed artifacts. Frontend source reorganization only.
